Use testfixtures
