+++ internal/executor/service/pod_issue_handler_test.go
@@ -4,11 +4,11 @@ import (
 	"testing"
 	"time"
 
+	"github.com/armadaproject/armada/internal/executor/reporter"
 	"github.com/stretchr/testify/assert"
 	v1 "k8s.io/api/core/v1"
 
 	fakecontext "github.com/armadaproject/armada/internal/executor/context/fake"
-	reporter_fake "github.com/armadaproject/armada/internal/executor/reporter/fake"
 	"github.com/armadaproject/armada/internal/executor/util"
 	"github.com/armadaproject/armada/pkg/api"
 )
@@ -102,9 +102,9 @@ func TestPodIssueService_ReportsFailed_IfDeletedExternally(t *testing.T) {
 	assert.Equal(t, failedEvent.JobId, util.ExtractJobId(runningPod))
 }
 
-func setupTestComponents() (*PodIssueService, *fakecontext.SyncFakeClusterContext, *reporter_fake.FakeEventReporter) {
+func setupTestComponents() (*PodIssueService, *fakecontext.SyncFakeClusterContext, *reporter.FakeEventReporter) {
 	fakeClusterContext := fakecontext.NewSyncFakeClusterContext()
-	eventReporter := reporter_fake.NewFakeEventReporter()
+	eventReporter := reporter.NewFakeEventReporter()
 	pendingPodChecker := makePodChecker()
 
 	podIssueHandler := NewPodIssueService(
